requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.2
wincertstore==0.2
WTForms==2.2.1
